Update interface.md
the previous description was for the Syncade interface -- I have removed it from here. However, we do need to come up with an introduction. We'll definitely have to bring in Dan Mabee to help with the content, but we could probably use the same structure that was laid out for Syncade: 
- first mention what versions of Emerson DeltaV BES we support
- talk about how this interface queries the EMDVB data, what this interface creates from the EMDVB data, and a bit about the resulting hierarchies created in PI and AF